Fix the bug in box_iou_rotated and add new version

+++ python/jdet/models/boxes/assigner.py
@@ -92,7 +92,7 @@ class MaxIoUAssigner:
             raise ValueError('No gt or bboxes')
 
         overlaps = self.iou_calculator(gt_bboxes, bboxes)
-
+        
         if (self.ignore_iof_thr > 0) and (gt_bboxes_ignore is not None) and (
                 gt_bboxes_ignore.numel() > 0):
             if self.ignore_wrt_candidates:
@@ -125,7 +125,7 @@ class MaxIoUAssigner:
         num_gts, num_bboxes = overlaps.size(0), overlaps.size(1)
 
         # 1. assign -1 by default
-        assigned_gt_inds = jt.full((num_bboxes,), -1).int()
+        assigned_gt_inds = jt.full((num_bboxes,), -1).long()
 
         # for each anchor, which gt best overlaps with it
         # for each anchor, the max iou of all gts
@@ -164,7 +164,7 @@ class MaxIoUAssigner:
                 assigned_labels[pos_inds] = gt_labels[assigned_gt_inds[pos_inds] - 1]
         else:
             assigned_labels = None
-            
+
         return AssignResult(num_gts, assigned_gt_inds, max_overlaps, labels=assigned_labels)
 
 @BOXES.register_module()
